src/helloservice/pyproject.toml
[tool.poetry]
name = "helloservice"
version = "0.1.0"
description = ""
authors = ["Matthew McLeod <mtthwmcld@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.35.0"
google-api-python-client = "^1.12.8"
grpcio-health-checking = "^1.35.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
grpcio-tools = "^1.35.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
